backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.18
python-multipart>=0.0.9
httpx>=0.27.0
openai>=1.30.0
fpdf2>=2.7.9
yt-dlp>=2024.5.1
python-dotenv>=1.0.0
